TIM.ENH: remove unused channels from OrbIntlk.
Orbit interlock is enabled only for sectors 6, 8, 10 and 14, so the high
level trigger doesn't need to control any other crate. This allows us to
disable the trigger for other crates, to avoid spurious events causing a
beam dump. We removed even the BPMs from those crates which aren't part
of the orbit interlock system, leaving only the BPMs for the straights.

This only became an issue when we enabled event reception (RxEnbl) in
the Fout devices for all timing AFCs. This was done in [1] with commit
656fb570 (orbintlk.ENH: fix Fout RxEnbl configuration to enable all
outputs, 2025-04-09).

Keeping the other SI BPMs for those sectors in the list was considered,
since they are all connected to the same trigger line and can therefore
affect it, but we decided against it, because by that logic we should
also include the BO BPMs and FOFB boards, which would make maintaing the
list much more complicated and error-prone.

[1] https://github.com/lnls-sirius/dev-packages