TIM.ENH: add OutOrbIntlk trigger.
This trigger includes the orbit interlock trigger line for all crates
which aren't included in the orbit interlock system. This way, it's
simple to guarantee they are configured correctly and consistently,
without requiring adding them to any machine configurations.

The lines with the crates included in the OrbIntlk trigger are commented
out to make it easier to review the trigger and match it against
OrbIntlk.