TIM.ENH: add OutOrbIntlk trigger.
This trigger includes the orbit interlock trigger line for all crates
which aren't included in the orbit interlock system. This way, it's
simple to guarantee they are configured correctly and consistently,
without requiring adding them to any machine configurations.

In order to make it clear that this trigger is a complement of OrbIntlk,
and to make it easier to review its definition by matching it against
OrbIntlk, we listed only the M1 and M2 BPMs for each crate, and
commented out the lines with the crates already covered by OrbIntlk.

The OutOrbIntlk trigger will be removed if orbit interlock ever includes
the BC dipoles, via the C2 and C3-1 BPMs, because in that situation all
crates will be in OrbIntlk.